# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
}

export interface Project {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  imageUrl: string;
  displayOrder: number;
  createdAt: string;
  /** @nullable */
  challenge?: string | null;
  /** @nullable */
  solution?: string | null;
  /** @nullable */
  results?: string | null;
  /** @nullable */
  technologies?: string[] | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  gallery?: string[] | null;
}

export interface ProjectInput {
  title: string;
  description?: string;
  category: string;
  imageUrl: string;
  displayOrder?: number;
  challenge?: string;
  solution?: string;
  results?: string;
  technologies?: string[];
  websiteUrl?: string;
  gallery?: string[];
}

export interface ProjectUpdate {
  title?: string;
  /** @nullable */
  description?: string | null;
  category?: string;
  imageUrl?: string;
  displayOrder?: number;
  /** @nullable */
  challenge?: string | null;
  /** @nullable */
  solution?: string | null;
  /** @nullable */
  results?: string | null;
  /** @nullable */
  technologies?: string[] | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  gallery?: string[] | null;
}

export interface Client {
  id: number;
  name: string;
  logoUrl: string;
  /** @nullable */
  website?: string | null;
  displayOrder: number;
  featured: boolean;
  createdAt: string;
}

export interface ClientInput {
  name: string;
  logoUrl: string;
  website?: string;
  displayOrder?: number;
  featured?: boolean;
}

export interface ClientUpdate {
  name?: string;
  logoUrl?: string;
  /** @nullable */
  website?: string | null;
  displayOrder?: number;
  featured?: boolean;
}

export interface ReorderInput {
  ids: number[];
}

export interface SiteSettings {
  phone: string;
  email: string;
  instagram: string;
  whatsapp: string;
  facebook: string;
  heroTitle: string;
  heroSubtitle: string;
  aboutText: string;
  logoUrl: string;
}

export interface SiteSettingsUpdate {
  phone?: string;
  email?: string;
  instagram?: string;
  whatsapp?: string;
  facebook?: string;
  heroTitle?: string;
  heroSubtitle?: string;
  aboutText?: string;
  logoUrl?: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}