# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: projects
    description: Portfolio projects
  - name: clients
    description: Client logos
  - name: settings
    description: Site settings
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/logout:
    post:
      operationId: authLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "204":
          description: Logged out
  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get current session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /projects:
    get:
      operationId: getProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/reorder:
    patch:
      operationId: reorderProjects
      tags: [projects]
      summary: Reorder projects (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: Clients list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/reorder:
    patch:
      operationId: reorderClients
      tags: [clients]
      summary: Reorder clients (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
  /clients/{id}:
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get site settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update site settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
        username:
          type: string
    Project:
      type: object
      required: [id, title, category, imageUrl, displayOrder, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        imageUrl:
          type: string
        displayOrder:
          type: integer
        createdAt:
          type: string
        challenge:
          type: ["string", "null"]
        solution:
          type: ["string", "null"]
        results:
          type: ["string", "null"]
        technologies:
          type: ["array", "null"]
          items:
            type: string
        websiteUrl:
          type: ["string", "null"]
        gallery:
          type: ["array", "null"]
          items:
            type: string
    ProjectInput:
      type: object
      required: [title, category, imageUrl]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        displayOrder:
          type: integer
        challenge:
          type: string
        solution:
          type: string
        results:
          type: string
        technologies:
          type: array
          items:
            type: string
        websiteUrl:
          type: string
        gallery:
          type: array
          items:
            type: string
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        imageUrl:
          type: string
        displayOrder:
          type: integer
        challenge:
          type: ["string", "null"]
        solution:
          type: ["string", "null"]
        results:
          type: ["string", "null"]
        technologies:
          type: ["array", "null"]
          items:
            type: string
        websiteUrl:
          type: ["string", "null"]
        gallery:
          type: ["array", "null"]
          items:
            type: string
    Client:
      type: object
      required: [id, name, logoUrl, displayOrder, featured, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: string
        website:
          type: ["string", "null"]
        displayOrder:
          type: integer
        featured:
          type: boolean
        createdAt:
          type: string
    ClientInput:
      type: object
      required: [name, logoUrl]
      properties:
        name:
          type: string
        logoUrl:
          type: string
        website:
          type: string
        displayOrder:
          type: integer
        featured:
          type: boolean
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
        website:
          type: ["string", "null"]
        displayOrder:
          type: integer
        featured:
          type: boolean
    ReorderInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer
    SiteSettings:
      type: object
      required: [phone, email, instagram, whatsapp, facebook, heroTitle, heroSubtitle, aboutText, logoUrl]
      properties:
        phone:
          type: string
        email:
          type: string
        instagram:
          type: string
        whatsapp:
          type: string
        facebook:
          type: string
        heroTitle:
          type: string
        heroSubtitle:
          type: string
        aboutText:
          type: string
        logoUrl:
          type: string
    SiteSettingsUpdate:
      type: object
      properties:
        phone:
          type: string
        email:
          type: string
        instagram:
          type: string
        whatsapp:
          type: string
        facebook:
          type: string
        heroTitle:
          type: string
        heroSubtitle:
          type: string
        aboutText:
          type: string
        logoUrl:
          type: string
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"