# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: number;
  name: string;
  logoUrl: string;
  /** @nullable */
  website?: string | null;
  displayOrder: number;
  featured: boolean;
  createdAt: string;
}