# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectInput {
  title: string;
  description?: string;
  category: string;
  imageUrl: string;
  displayOrder?: number;
  challenge?: string;
  solution?: string;
  results?: string;
  technologies?: string[];
  websiteUrl?: string;
  gallery?: string[];
}