# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  title?: string;
  /** @nullable */
  description?: string | null;
  category?: string;
  imageUrl?: string;
  displayOrder?: number;
  /** @nullable */
  challenge?: string | null;
  /** @nullable */
  solution?: string | null;
  /** @nullable */
  results?: string | null;
  /** @nullable */
  technologies?: string[] | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  gallery?: string[] | null;
}